requirements.txt
django
PyPDF2
google-generativeai